requirements.txt
asgiref==3.8.1
channels==4.1.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
paho-mqtt==2.1.0
pvporcupine==3.0.2
PyAudio==0.2.14
python-dotenv==1.0.1
sqlparse==0.5.0
